+++ com.unity.toonshader/CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Fixed:
 * HDRP:reduced shader variants by disabling some debug pragmas.
 * URP:UTS materials don't recieve shadows when using newer than Unity 2021.1.
-
+* Legacy:Unable to disable Outline.
 ### Updated
 * More effective and strict internal tests.
 ## [0.4.1-preview] - 2021-10-20